refactor to get rid of unneeded components
