+++ packages/frontend-v2/components/Form/InputGroup.tsx
@@ -6,7 +6,6 @@ export const InputGroup: React.FC = ({ children }) => (
     justifyContent="space-evenly"
     alignItems="center"
     width="100%"
-    height="80%"
     minHeight="3xs"
   >
     {children}
