+++ packages/frontend-v2/components/Form/index.ts
@@ -1,5 +1,3 @@
-export * from "./LoginForm";
 export * from './Input'
 export * from './InputGroup'
-export * from "./SignUpForm";
 export * from './FormSections'
