--- packages/frontend-v2/components/Header/SignIn.tsx
@@ -1,11 +1,0 @@
-import { Flex, Text } from "@chakra-ui/react";
-
-
-export const SignIn: React.FC = () => (
-  <Flex>
-    Image
-    <Text fontSize='md'>
-      Login
-    </Text>
-  </Flex>
-)
