+++ packages/frontend-v2/components/Header/index.ts
@@ -1,3 +1,2 @@
 export { Logo } from "./Logo";
 export { HeaderContainer } from "./HeaderContainer";
-export { SignIn } from './SignIn'
